# openapi.yaml
components:
  schemas:
...
    Folder:
      properties:
        messages:
          type: integer
        name:
          type: string
        uid_next:
          type: integer
        uid_validity:
          type: integer
      type: object
    Mailbox:
      properties:
        api_base:
          type: string
        created:
          format: date-time
          type: string
        id:
          type: string
        name:
          type: string
        recipients:
          items:
            type: string
          type: array
        username:
          type: string
      type: object
...
  securitySchemes:
    basicAuth:
      description: Optional configured HTTP credentials; the send API can use separate credentials.
      scheme: basic
      type: http
info:
  description: Independent shared SMTP, TLS IMAP and HTTP test mailbox. No browser UI or POP3. Message-scoped toxics isolate concurrent test traffic. Outbound mail is disabled unless relay/forwarding is configured. Default memory storage; optional bbolt persistence. Provision independent accounts at /api/v1/mailboxes and prefix standard routes with the returned api_base. Root routes address the default mailbox only. SMTP recipient routing selects the mailbox at RCPT TO; sender-stage toxics are therefore evaluated at the first RCPT, not MAIL FROM.
  title: Mail Sandbox
  version: 2.1.0
openapi: 3.1.0
paths:
...
  /api/events:
    get:
      description: 'Text JSON {Type: new|update|delete, Data: summary or {ID}}. Best effort, bounded queue; slow consumers disconnect.'
      operationId: getApiEvents
      responses:
        "101":
          description: Success
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
        "401":
          content:
            text/plain:
              schema:
                type: string
          description: Configured authentication required
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Resource not found
      summary: Upgrade to WebSocket for mailbox events
  /api/v1/chaos:
    get:
      operationId: getApiV1Chaos
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  Enabled:
                    type: boolean
                  MessageScoped:
                    type: boolean
                  Toxics:
                    type: string
                type: object
          description: Success
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
        "401":
          content:
            text/plain:
              schema:
                type: string
          description: Configured authentication required
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Resource not found
      summary: Report global chaos disabled
    put:
      description: Use /api/v1/toxics with an exact message selector. Global chaos violates concurrency isolation.
      operationId: putApiV1Chaos
      responses:
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
        "401":
          content:
            text/plain:
              schema:
                type: string
          description: Configured authentication required
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Resource not found
      summary: Reject global failure injection
  /api/v1/folders:
    get:
      operationId: getApiV1Folders
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Folder'
                type: array
          description: Success
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
        "401":
          content:
            text/plain:
              schema:
                type: string
          description: Configured authentication required
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Resource not found
      summary: List folders and counters
    post:
      operationId: postApiV1Folders
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              properties:
                name:
                  type: string
              required:
                - name
              type: object
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                properties:
                  name:
                    type: string
                type: object
          description: Success
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
        "401":
          content:
            text/plain:
              schema:
                type: string
          description: Configured authentication required
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Resource not found
      summary: Create a folder
  /api/v1/info:
    get:
      operationId: getApiV1Info
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: Success
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
        "401":
          content:
            text/plain:
              schema:
                type: string
          description: Configured authentication required
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Resource not found
      summary: Runtime/store statistics
  /api/v1/mailboxes:
    get:
      operationId: getApiV1Mailboxes
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Mailbox'
                type: array
          description: Success
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
        "401":
          content:
            text/plain:
              schema:
                type: string
          description: Configured authentication required
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Resource not found
      summary: List independent mailbox accounts (no passwords)
    post:
      description: Password is returned only on creation. Each mailbox has isolated mail, folders, flags/tags, retention, toxics and notifications. SMTP credentials remain server-wide; IMAP credentials select the account. Bindings must not overlap. SMTP transactions cannot mix accounts.
      operationId: postApiV1Mailboxes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MailboxCreation'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedMailbox'
          description: Success
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
        "401":
          content:
            text/plain:
              schema:
                type: string
          description: Configured authentication required
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Resource not found
      summary: Provision an independent mailbox account
  /api/v1/mailboxes/{mailbox}:
    delete:
      description: The default mailbox cannot be deleted. Existing sessions cannot read deleted mail. Deleted recipient bindings reject SMTP until explicitly reassigned; they do not fall through to default. With persistence enabled, messages and catalog metadata are removed from their files.
      operationId: deleteApiV1MailboxesMailbox
      parameters:
        - in: path
          name: mailbox
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Success
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
        "401":
          content:
            text/plain:
              schema:
                type: string
          description: Configured authentication required
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Resource not found
      summary: Delete a mailbox and its data
    get:
      operationId: getApiV1MailboxesMailbox
      parameters:
        - in: path
          name: mailbox
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mailbox'
          description: Success
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
        "401":
          content:
            text/plain:
              schema:
                type: string
          description: Configured authentication required
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Resource not found
      summary: Get mailbox metadata without credentials
  /api/v1/message/{id}:
    get:
      description: id is a database UUID or latest. Inspection endpoints other than detail do not mark read.
      operationId: getApiV1MessageId
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: Success
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
        "401":
          content:
            text/plain:
              schema:
                type: string
          description: Configured authentication required
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Resource not found
      summary: Get message detail and mark it read
  /api/v1/message/{id}/headers:
    get:
      description: id is a database UUID or latest. Inspection endpoints other than detail do not mark read.
      operationId: getApiV1MessageIdHeaders
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    type: string
                  type: array
                type: object
          description: Success
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
        "401":
          content:
            text/plain:
              schema:
                type: string
          description: Configured authentication required
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Resource not found
      summary: Get decoded headers
  /api/v1/message/{id}/html-check:
    get:
      description: id is a database UUID or latest. Inspection endpoints other than detail do not mark read.
      operationId: getApiV1MessageIdHtmlCheck
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: Success
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
        "401":
          content:
            text/plain:
              schema:
                type: string
          description: Configured authentication required
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Resource not found
      summary: Analyze HTML/CSS compatibility
  /api/v1/message/{id}/link-check:
    get:
      description: id is a database UUID or latest. Inspection endpoints other than detail do not mark read.
      operationId: getApiV1MessageIdLinkCheck
      parameters:
        - description: ""
          in: query
          name: follow
          schema:
            default: false
            type: boolean
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  Errors:
                    type: integer
                  Links:
                    items:
                      properties:
                        Status:
                          type: string
                        StatusCode:
                          type: integer
                        URL:
                          type: string
                      type: object
                    type: array
                type: object
          description: Success
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
        "401":
          content:
            text/plain:
              schema:
                type: string
          description: Configured authentication required
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: Resource not found
      summary: Check links
...
  /mailboxes/{mailbox}/:
    get:
      operationId: mailboxGet
      parameters:
        - description: Mailbox ID from provisioning, or default. Every data and toxic operation stays within this scope.
          in: path
          name: mailbox
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: Success
      summary: Service metadata
...
security:
  - {}
  - basicAuth: []
servers:
  - url: http://localhost:8026